invoices checkbox fix
